09-serverless/pyproject.toml
[project]
name = "09-serverless"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy",
    "pillow",
    "onnxruntime"
]